+++ Sources/Sentry/include/SentryScope+Private.h
@@ -25,6 +25,21 @@ NS_ASSUME_NONNULL_BEGIN
  */
 @property (nonatomic, strong) SentryPropagationContext *propagationContext;
 
+/**
+ * Set global extra -> these will be sent with every event
+ */
+@property (atomic, strong) NSMutableDictionary<NSString *, id> *extraDictionary;
+
+/**
+ * Set the fingerprint of an event to determine the grouping
+ */
+@property (atomic, strong) NSMutableArray<NSString *> *fingerprintArray;
+
+/**
+ * SentryLevel of the event
+ */
+@property (atomic) enum SentryLevel levelEnum;
+
 @property (nonatomic, nullable, copy) NSString *currentScreen;
 
 - (NSArray<SentryBreadcrumb *> *)breadcrumbs;