+++ Tests/SentryTests/Persistence/SentryScopePersistentStoreTests.swift
@@ -312,18 +312,263 @@ class SentryScopePersistentStoreTests: XCTestCase {
         XCTAssertEqual(decodedUser.username, "testuser")
     }
 
+    // MARK: - Level Tests
+
+    func testReadPreviousLevelFromDisk_whenValidDataInPreviousLevelFile_shouldReturnDecodedLevel() throws {
+        // -- Arrange --
+        let fm = FileManager.default
+        let levelData = Data("\(SentryLevel.fatal.rawValue)".utf8)
+        let previousLevelFileURL = sut.previousFileURLFor(field: .level)
+        try levelData.write(to: previousLevelFileURL)
+        XCTAssertTrue(fm.fileExists(atPath: previousLevelFileURL.path))
+
+        // -- Act --
+        let result = sut.readPreviousLevelFromDisk()
+
+        // -- Assert --
+        XCTAssertEqual(result, .fatal)
+    }
+
+    func testReadPreviousLevelFromDisk_whenInvalidDataInPreviousLevelFile_shouldReturnNoneLevel() throws {
+        // -- Arrange --
+        let fm = FileManager.default
+        let levelData = Data("invalid".utf8)
+        let previousLevelFileURL = sut.previousFileURLFor(field: .level)
+        try levelData.write(to: previousLevelFileURL)
+        XCTAssertTrue(fm.fileExists(atPath: previousLevelFileURL.path))
+
+        // -- Act --
+        let result = sut.readPreviousLevelFromDisk()
+
+        // -- Assert --
+        XCTAssertEqual(result, .none)
+    }
+
+    func testReadPreviousLevelFromDisk_whenPreviousLevelFileUnavailable_shouldReturnNoneLevel() throws {
+        // -- Arrange --
+        let fm = FileManager.default
+        let previousLevelFileURL = sut.previousFileURLFor(field: .level)
+        if fm.fileExists(atPath: previousLevelFileURL.path) {
+            try fm.removeItem(at: previousLevelFileURL)
+        }
+        XCTAssertFalse(fm.fileExists(atPath: previousLevelFileURL.path))
+
+        // -- Act --
+        let result = sut.readPreviousLevelFromDisk()
+
+        // -- Assert --
+        XCTAssertEqual(result, .none)
+    }
+
+    func testWriteLevelToDisk_whenValidLevel_shouldWriteToLevelFile() throws {
+        // -- Arrange --
+        let fm = FileManager.default
+        let level = NSNumber(value: SentryLevel.fatal.rawValue)
+        let levelFileURL = sut.currentFileURLFor(field: .level)
+
+        // Check pre-conditions
+        XCTAssertFalse(fm.fileExists(atPath: levelFileURL.path))
+
+        // -- Act --
+        sut.writeLevelToDisk(level: level)
+
+        // -- Assert --
+        XCTAssertTrue(fm.fileExists(atPath: levelFileURL.path))
+        
+        let writtenData = try Data(contentsOf: levelFileURL)
+        let writtenString = String(data: writtenData, encoding: .utf8)
+        XCTAssertEqual(writtenString, "\(SentryLevel.fatal.rawValue)")
+    }
+
+    // MARK: - Extras Tests
+
+    func testReadPreviousExtrasFromDisk_whenValidJSONInPreviousExtrasFile_shouldReturnDecodedData() throws {
+        // -- Arrange --
+        let fm = FileManager.default
+        let extras: [String: Any] = ["key1": "value1", "key2": 42, "key3": true]
+        let extrasData = try XCTUnwrap(SentrySerialization.data(withJSONObject: extras))
+        let previousExtrasFileURL = sut.previousFileURLFor(field: .extras)
+        try extrasData.write(to: previousExtrasFileURL)
+        XCTAssertTrue(fm.fileExists(atPath: previousExtrasFileURL.path))
+
+        // -- Act --
+        let result = try XCTUnwrap(sut.readPreviousExtrasFromDisk())
+
+        // -- Assert --
+        XCTAssertEqual(result.count, 3)
+        XCTAssertEqual(result["key1"] as? String, "value1")
+        XCTAssertEqual(result["key2"] as? Int, 42)
+        XCTAssertEqual(result["key3"] as? Bool, true)
+    }
+
+    func testReadPreviousExtrasFromDisk_whenInvalidJSONInPreviousExtrasFile_shouldReturnNil() throws {
+        // -- Arrange --
+        let fm = FileManager.default
+        let data = Data("<TEST DATA>".utf8)
+        let previousExtrasFileURL = sut.previousFileURLFor(field: .extras)
+        try data.write(to: previousExtrasFileURL)
+        XCTAssertTrue(fm.fileExists(atPath: previousExtrasFileURL.path))
+
+        // -- Act --
+        let result = sut.readPreviousExtrasFromDisk()
+
+        // -- Assert --
+        XCTAssertNil(result)
+    }
+
+    func testReadPreviousExtrasFromDisk_whenPreviousExtrasFileUnavailable_shouldReturnNil() throws {
+        // -- Arrange --
+        let fm = FileManager.default
+        let previousExtrasFileURL = sut.previousFileURLFor(field: .extras)
+        if fm.fileExists(atPath: previousExtrasFileURL.path) {
+            try fm.removeItem(at: previousExtrasFileURL)
+        }
+        XCTAssertFalse(fm.fileExists(atPath: previousExtrasFileURL.path))
+
+        // -- Act --
+        let result = sut.readPreviousExtrasFromDisk()
+
+        // -- Assert --
+        XCTAssertNil(result)
+    }
+
+    func testWriteExtrasToDisk_whenValidExtras_shouldWriteToExtrasFile() throws {
+        // -- Arrange --
+        let fm = FileManager.default
+        let extras: [String: Any] = ["key1": "value1", "key2": 42, "key3": true]
+        let extrasFileURL = sut.currentFileURLFor(field: .extras)
+
+        // Check pre-conditions
+        XCTAssertFalse(fm.fileExists(atPath: extrasFileURL.path))
+
+        // -- Act --
+        sut.writeExtrasToDisk(extras: extras)
+
+        // -- Assert --
+        XCTAssertTrue(fm.fileExists(atPath: extrasFileURL.path))
+        
+        let writtenData = try Data(contentsOf: extrasFileURL)
+        let serializedData = try XCTUnwrap(SentrySerialization.deserializeDictionary(fromJsonData: writtenData))
+        
+        XCTAssertEqual(serializedData.count, 3)
+        XCTAssertEqual(serializedData["key1"] as? String, "value1")
+        XCTAssertEqual(serializedData["key2"] as? Int, 42)
+        XCTAssertEqual(serializedData["key3"] as? Bool, true)
+    }
+
+    func testWriteExtrasToDisk_whenInvalidJSONExtras_shouldNotWriteToExtrasFile() throws {
+        // -- Arrange --
+        let fm = FileManager.default
+        let extras: [String: Any] = ["key1": Double.infinity]
+        let extrasFileURL = sut.currentFileURLFor(field: .extras)
+        if fm.fileExists(atPath: extrasFileURL.path) {
+            try fm.removeItem(at: extrasFileURL)
+        }
+
+        // Check pre-conditions
+        XCTAssertFalse(fm.fileExists(atPath: extrasFileURL.path))
+
+        // -- Act --
+        sut.writeExtrasToDisk(extras: extras)
+
+        // -- Assert --
+        XCTAssertFalse(fm.fileExists(atPath: extrasFileURL.path))
+    }
+
+    // MARK: - Fingerprint Tests
+
+    func testReadPreviousFingerprintFromDisk_whenValidJSONInPreviousFingerprintFile_shouldReturnDecodedData() throws {
+        // -- Arrange --
+        let fm = FileManager.default
+        let fingerprint = ["fp1", "fp2", "fp3"]
+        let fingerprintData = try XCTUnwrap(SentrySerialization.data(withJSONObject: fingerprint))
+        let previousFingerprintFileURL = sut.previousFileURLFor(field: .fingerprint)
+        try fingerprintData.write(to: previousFingerprintFileURL)
+        XCTAssertTrue(fm.fileExists(atPath: previousFingerprintFileURL.path))
+
+        // -- Act --
+        let result = try XCTUnwrap(sut.readPreviousFingerprintFromDisk())
+
+        // -- Assert --
+        XCTAssertEqual(result, fingerprint)
+    }
+
+    func testReadPreviousFingerprintFromDisk_whenInvalidJSONInPreviousFingerprintFile_shouldReturnNil() throws {
+        // -- Arrange --
+        let fm = FileManager.default
+        let data = Data("<TEST DATA>".utf8)
+        let previousFingerprintFileURL = sut.previousFileURLFor(field: .fingerprint)
+        try data.write(to: previousFingerprintFileURL)
+        XCTAssertTrue(fm.fileExists(atPath: previousFingerprintFileURL.path))
+
+        // -- Act --
+        let result = sut.readPreviousFingerprintFromDisk()
+
+        // -- Assert --
+        XCTAssertNil(result)
+    }
+
+    func testReadPreviousFingerprintFromDisk_whenPreviousFingerprintFileUnavailable_shouldReturnNil() throws {
+        // -- Arrange --
+        let fm = FileManager.default
+        let previousFingerprintFileURL = sut.previousFileURLFor(field: .fingerprint)
+        if fm.fileExists(atPath: previousFingerprintFileURL.path) {
+            try fm.removeItem(at: previousFingerprintFileURL)
+        }
+        XCTAssertFalse(fm.fileExists(atPath: previousFingerprintFileURL.path))
+
+        // -- Act --
+        let result = sut.readPreviousFingerprintFromDisk()
+
+        // -- Assert --
+        XCTAssertNil(result)
+    }
+
+    func testWriteFingerprintToDisk_whenValidFingerprint_shouldWriteToFingerprintFile() throws {
+        // -- Arrange --
+        let fm = FileManager.default
+        let fingerprint = ["fp1", "fp2", "fp3"]
+        let fingerprintFileURL = sut.currentFileURLFor(field: .fingerprint)
+
+        // Check pre-conditions
+        XCTAssertFalse(fm.fileExists(atPath: fingerprintFileURL.path))
+
+        // -- Act --
+        sut.writeFingerprintToDisk(fingerprint: fingerprint)
+
+        // -- Assert --
+        XCTAssertTrue(fm.fileExists(atPath: fingerprintFileURL.path))
+        
+        let writtenData = try Data(contentsOf: fingerprintFileURL)
+        let serializedData = try XCTUnwrap(SentrySerialization.deserializeArray(fromJsonData: writtenData))
+        
+        XCTAssertEqual(serializedData.count, 3)
+        XCTAssertEqual(serializedData[0] as? String, "fp1")
+        XCTAssertEqual(serializedData[1] as? String, "fp2")
+        XCTAssertEqual(serializedData[2] as? String, "fp3")
+    }
+
     func testDeleteAllCurrentState_shouldDeleteAllCurrentFiles() throws {
         // -- Arrange --
         let fm = FileManager.default
         let contextFileURL = sut.currentFileURLFor(field: .context)
         let userFileURL = sut.currentFileURLFor(field: .user)
+        let levelFileURL = sut.currentFileURLFor(field: .level)
+        let extrasFileURL = sut.currentFileURLFor(field: .extras)
+        let fingerprintFileURL = sut.currentFileURLFor(field: .fingerprint)
         
         // Create test files
         try "context data".write(to: contextFileURL, atomically: true, encoding: .utf8)
         try "user data".write(to: userFileURL, atomically: true, encoding: .utf8)
+        try "level data".write(to: levelFileURL, atomically: true, encoding: .utf8)
+        try "extras data".write(to: extrasFileURL, atomically: true, encoding: .utf8)
+        try "fingerprint data".write(to: fingerprintFileURL, atomically: true, encoding: .utf8)
         
         XCTAssertTrue(fm.fileExists(atPath: contextFileURL.path))
         XCTAssertTrue(fm.fileExists(atPath: userFileURL.path))
+        XCTAssertTrue(fm.fileExists(atPath: levelFileURL.path))
+        XCTAssertTrue(fm.fileExists(atPath: extrasFileURL.path))
+        XCTAssertTrue(fm.fileExists(atPath: fingerprintFileURL.path))
 
         // -- Act --
         sut.deleteAllCurrentState()
@@ -331,6 +576,9 @@ class SentryScopePersistentStoreTests: XCTestCase {
         // -- Assert --
         XCTAssertFalse(fm.fileExists(atPath: contextFileURL.path))
         XCTAssertFalse(fm.fileExists(atPath: userFileURL.path))
+        XCTAssertFalse(fm.fileExists(atPath: levelFileURL.path))
+        XCTAssertFalse(fm.fileExists(atPath: extrasFileURL.path))
+        XCTAssertFalse(fm.fileExists(atPath: fingerprintFileURL.path))
     }
 
     func testDeleteAllPreviousState_shouldDeleteAllPreviousFiles() throws {
@@ -338,13 +586,22 @@ class SentryScopePersistentStoreTests: XCTestCase {
         let fm = FileManager.default
         let previousContextFileURL = sut.previousFileURLFor(field: .context)
         let previousUserFileURL = sut.previousFileURLFor(field: .user)
+        let previousLevelFileURL = sut.previousFileURLFor(field: .level)
+        let previousExtrasFileURL = sut.previousFileURLFor(field: .extras)
+        let previousFingerprintFileURL = sut.previousFileURLFor(field: .fingerprint)
         
         // Create test files
         try "previous context data".write(to: previousContextFileURL, atomically: true, encoding: .utf8)
         try "previous user data".write(to: previousUserFileURL, atomically: true, encoding: .utf8)
+        try "previous level data".write(to: previousLevelFileURL, atomically: true, encoding: .utf8)
+        try "previous extras data".write(to: previousExtrasFileURL, atomically: true, encoding: .utf8)
+        try "previous fingerprint data".write(to: previousFingerprintFileURL, atomically: true, encoding: .utf8)
         
         XCTAssertTrue(fm.fileExists(atPath: previousContextFileURL.path))
         XCTAssertTrue(fm.fileExists(atPath: previousUserFileURL.path))
+        XCTAssertTrue(fm.fileExists(atPath: previousLevelFileURL.path))
+        XCTAssertTrue(fm.fileExists(atPath: previousExtrasFileURL.path))
+        XCTAssertTrue(fm.fileExists(atPath: previousFingerprintFileURL.path))
 
         // -- Act --
         sut.deleteAllPreviousState()
@@ -352,5 +609,8 @@ class SentryScopePersistentStoreTests: XCTestCase {
         // -- Assert --
         XCTAssertFalse(fm.fileExists(atPath: previousContextFileURL.path))
         XCTAssertFalse(fm.fileExists(atPath: previousUserFileURL.path))
+        XCTAssertFalse(fm.fileExists(atPath: previousLevelFileURL.path))
+        XCTAssertFalse(fm.fileExists(atPath: previousExtrasFileURL.path))
+        XCTAssertFalse(fm.fileExists(atPath: previousFingerprintFileURL.path))
     }
 }