+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Removed `APPLICATION_EXTENSION_API_ONLY` requirement (#5524)
 - Improve launch profile configuration management (#5318)
 - Record user for watchdog termination events (#5558)
+- Add support for extras, fingerprint, and level watchdog termination events (#5569)
 
 ## 8.53.1